refactor: change tests folder structure
