+++ jest.config.ts
@@ -1,6 +1,5 @@
 module.exports = {
   collectCoverage: true,
-  // on node 14.x coverage provider v8 offers good speed and more or less good report
   coverageProvider: 'v8',
   collectCoverageFrom: [
     '**/pages/**/*.{js,jsx,ts,tsx}',
@@ -13,9 +12,6 @@ module.exports = {
     '!pages/_document.tsx',
   ],
   moduleNameMapper: {
-    // '/.(css|less|scss|sass)$/': 'identity-obj-proxy',
-    // Handle CSS imports (with CSS modules)
-    // https://jestjs.io/docs/webpack#mocking-css-modules
     '^.+\\.module\\.(css|sass|scss)$': 'identity-obj-proxy',
 
     '^.+\\.(css|sass|scss)$': '<rootDir>/__mocks__/styleMock.js',
@@ -24,13 +20,9 @@ module.exports = {
 
     '^@/components/(.*)$': '<rootDir>/components/$1',
   },
-  // Add more setup options before each test is run
-  // setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/.next/'],
   testEnvironment: 'jest-environment-jsdom',
   transform: {
-    // Use babel-jest to transpile tests with the next/babel preset
-    // https://jestjs.io/docs/configuration#transform-objectstring-pathtotransformer--pathtotransformer-object
     '^.+\\.(js|jsx|ts|tsx)$': ['babel-jest', { presets: ['next/babel'] }],
   },
   transformIgnorePatterns: [
